Introduce FormatterTextRenderer
