+++ packages/flame/lib/src/text/common/glyph_info.dart
@@ -1,4 +1,3 @@
-
 /// Helper class that stores dimensions of a single glyph for a spritesheet-
 /// based font.
 class GlyphInfo {
