+++ packages/flame/lib/src/text/sprite_font_renderer.dart
@@ -1,10 +1,9 @@
 import 'dart:ui';
 
-import 'package:flame/src/anchor.dart';
 import 'package:flame/src/text/common/glyph_data.dart';
+import 'package:flame/src/text/formatter_text_renderer.dart';
 import 'package:flame/src/text/formatters/sprite_font_text_formatter.dart';
 import 'package:flame/src/text/text_renderer.dart';
-import 'package:vector_math/vector_math_64.dart';
 
 /// [TextRenderer] implementation that uses a spritesheet of various font glyphs
 /// to render text.
@@ -15,15 +14,16 @@ import 'package:vector_math/vector_math_64.dart';
 ///
 /// Currently, this class supports monospace fonts only -- the widths and the
 /// heights of all characters must be the same.
-/// Extra space between letters can be added via the [letterSpacing] parameter
+/// Extra space between letters can be added via the `letterSpacing` parameter
 /// (it can also be negative to "squish" characters together). Finally, the
-/// [scale] parameter allows scaling the font to be bigger/smaller relative to
+/// `scale` parameter allows scaling the font to be bigger/smaller relative to
 /// its size in the source image.
 ///
-/// The [paint] parameter is used to composite the character images onto the
+/// The `paint` parameter is used to composite the character images onto the
 /// canvas. Its default value will draw the character images as-is. Changing
 /// the opacity of the paint's color will make the text semi-transparent.
-class SpriteFontRenderer extends TextRenderer {
+class SpriteFontRenderer
+    extends FormatterTextRenderer<SpriteFontTextFormatter> {
   SpriteFontRenderer({
     required Image source,
     required double charWidth,
@@ -31,35 +31,14 @@ class SpriteFontRenderer extends TextRenderer {
     required Map<String, GlyphData> glyphs,
     double scale = 1,
     double letterSpacing = 0,
-  }) : formatter = SpriteFontTextFormatter(
-          source: source,
-          charWidth: charWidth,
-          charHeight: charHeight,
-          glyphs: glyphs,
-          scale: scale,
-          letterSpacing: letterSpacing,
+  }) : super(
+          SpriteFontTextFormatter(
+            source: source,
+            charWidth: charWidth,
+            charHeight: charHeight,
+            glyphs: glyphs,
+            scale: scale,
+            letterSpacing: letterSpacing,
+          ),
         );
-
-  final SpriteFontTextFormatter formatter;
-
-  @override
-  Vector2 measureText(String text) {
-    final box = formatter.format(text).metrics;
-    return Vector2(box.width, box.height);
-  }
-
-  @override
-  void render(
-    Canvas canvas,
-    String text,
-    Vector2 position, {
-    Anchor anchor = Anchor.topLeft,
-  }) {
-    final txt = formatter.format(text);
-    txt.translate(
-      position.x - txt.metrics.width * anchor.x,
-      position.y - txt.metrics.height * anchor.y - txt.metrics.top,
-    );
-    txt.render(canvas);
-  }
 }
